Reorganized enums
